Do not decode empty text packets (Fixes #269)
The commit bcd1a42f86aad12460b6eb836a5c317db55cba77 introduced the
feature to support binary packets with zero length. However, the
change introduced a bug which is fixed with this commit.

When receiving an empty packet which is not bytes (e.g. ''), the Packet class
tries to decode this (it did not previous the mentioned commit) and
throws an exception as it cannot read the first byte (which is supposed to
contain the packet type).

This behavior can occur when establishing a WebSocket connection. After the
PING packet has been sent (client.py#431), a PONG packet is expected to be
received (client.py#438). However, the WebSocket receive function can
return an empty string which now makes the packet decoding crash.
(see https://github.com/websocket-client/websocket-client/blob/d7aa5c9c244e227041764c63949d3121f84145ac/websocket/_core.py#L353-L368)
As this exception is not caught, it prevents further connection attempts.